feat: Add System.Numerics support and extend vector functionalities
- Integrated System.Numerics in various physics and graphics files for enhanced vector operations.
- Introduced VectorExtensions for common vector operations, including distance calculations, normalization, and quaternion manipulations.
- Added ManorGamePresets for configuring boid behaviors in the Manor simulation.
- Updated multiple demo scenes to utilize new vector functionalities and improve code clarity.
- Enhanced UI elements in ManorDemoManager for better user interaction.

+++ GPGames/GPGems.Visualization/ManorGameDemos/ManorDemoManager.xaml.cs
@@ -5,7 +5,8 @@ using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Shapes;
 using System.Windows.Threading;
-
+using System.Numerics;
+using GPGems.Core.Math;
 namespace GPGems.Visualization.ManorGameDemos
 {
     public partial class ManorDemoManager : UserControl
@@ -98,7 +99,7 @@ namespace GPGems.Visualization.ManorGameDemos
         private void ResetCurrentScene()
         {
             _frame = 0;
-            _currentScene?.Reset((int)SliderCount.Value, SliderSpeed.Value);
+            _currentScene?.Reset((int)SliderCount.Value, (float)SliderSpeed.Value);
             Render();
         }
